requirements.txt
click>=8.1.0
pypdf>=3.0.0
pyyaml>=6.0